Делим код на модули
